@@ -38,13 +38,13 @@ setup(
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
         'Operating System :: OS Independent',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
         'Topic :: Multimedia :: Video :: Conversion',
         'Topic :: Utilities',
     ],
     keywords='wrapper',
-    python_requires='>=3.6',
+    python_requires='>=3.9',
     use_scm_version=True,
     setup_requires=setup_requires,
     install_requires=install_requires,
