@@ -81,7 +81,7 @@ class MKVAttachment:
     def file_path(self, file_path):
         file_path = expanduser(file_path)
         if not isfile(file_path):
-            raise FileNotFoundError('"{}" does not exist'.format(file_path))
+            raise FileNotFoundError(f'"{file_path}" does not exist')
         self.mime_type = guess_type(file_path)[0]
         self.name = None
         self._file_path = file_path
