@@ -20,9 +20,7 @@ def verify_mkvmerge(mkvmerge_path='mkvmerge'):
         output = sp.check_output([mkvmerge_path, '-V']).decode()
     except (sp.CalledProcessError, FileNotFoundError):
         return False
-    if match('mkvmerge.*', output):
-        return True
-    return False
+    return bool(match('mkvmerge.*', output))
 
 
 def verify_matroska(file_path, mkvmerge_path='mkvmerge'):
@@ -34,19 +32,20 @@ def verify_matroska(file_path, mkvmerge_path='mkvmerge'):
         Alternate path to mkvmerge if it is not already in the $PATH variable.
     """
     if not verify_mkvmerge(mkvmerge_path=mkvmerge_path):
-        raise FileNotFoundError('mkvmerge is not at the specified path, add it there or change the mkvmerge_path '
-                                'property')
+        raise FileNotFoundError('mkvmerge is not at the specified path, add it there or change the mkvmerge_path property')
+
     if isinstance(file_path, os.PathLike):
         file_path = str(file_path)
     elif not isinstance(file_path, str):
-        raise TypeError('"{}" is not of type str'.format(file_path))
+        raise TypeError(f'"{file_path}" is not of type str')
     file_path = expanduser(file_path)
     if not isfile(file_path):
-        raise FileNotFoundError('"{}" does not exist'.format(file_path))
+        raise FileNotFoundError(f'"{file_path}" does not exist')
     try:
         info_json = json.loads(sp.check_output([mkvmerge_path, '-J', expanduser(file_path)]).decode())
-    except sp.CalledProcessError:
-        raise ValueError('"{}" could not be opened'.format(file_path))
+
+    except sp.CalledProcessError as e:
+        raise ValueError(f'"{file_path}" could not be opened') from e
     return info_json['container']['type'] == 'Matroska'
 
 
@@ -61,15 +60,17 @@ def verify_recognized(file_path, mkvmerge_path='mkvmerge'):
     if not verify_mkvmerge(mkvmerge_path=mkvmerge_path):
         raise FileNotFoundError('mkvmerge is not at the specified path, add it there or change the mkvmerge_path '
                                 'property')
+
     if not isinstance(file_path, str):
-        raise TypeError('"{}" is not of type str'.format(file_path))
+        raise TypeError(f'"{file_path}" is not of type str')
     file_path = expanduser(file_path)
     if not isfile(file_path):
-        raise FileNotFoundError('"{}" does not exist'.format(file_path))
+        raise FileNotFoundError(f'"{file_path}" does not exist')
     try:
         info_json = json.loads(sp.check_output([mkvmerge_path, '-J', file_path]).decode())
-    except sp.CalledProcessError:
-        raise ValueError('"{}" could not be opened'.format(file_path))
+
+    except sp.CalledProcessError as e:
+        raise ValueError(f'"{file_path}" could not be opened') from e
     return info_json['container']['recognized']
 
 
