@@ -5,11 +5,9 @@
 
 import bcp47
 
-def is_BCP47(language_ietf):
+
+def is_bcp47(language_ietf):
     if language_ietf == "und":
         return True
     else:
-        if language_ietf in bcp47.languages.values():
-            return True
-        else:
-            return False
+        return language_ietf in bcp47.languages.values()
